requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pyrtlsdr>=0.2.9
matplotlib>=3.7.0
bitstring>=4.0.0
PyQt6>=6.6.0
PyQt6-Charts>=6.6.0
sounddevice>=0.4.6
pyusb>=1.2.1